restore original shhs
