+++ docs/api/datasets/pyhealth.datasets.SHHSDataset.rst
@@ -1,9 +1,15 @@
 ﻿pyhealth.datasets.SHHSDataset
 ===================================
 
-The Sleep Heart Health Study (SHHS) dataset, refer to `doc <https://sleepdata.org/datasets/shhs>`_ for more information.
+The open Sleep-EDF Database Expanded database, refer to `doc <https://sleepdata.org/datasets/shhs>`_ for more information. 
 
 .. autoclass:: pyhealth.datasets.SHHSDataset
     :members:
     :undoc-members:
     :show-inheritance:
+
+   
+
+   
+   
+
