+++ docs/api/datasets/pyhealth.datasets.SHHSECGDataset.rst
@@ -1,0 +1,9 @@
+pyhealth.datasets.SHHSECGDataset
+=======================================
+
+The SHHS ECG dataset (ECG-focused loader for the Sleep Heart Health Study), refer to `doc <https://sleepdata.org/datasets/shhs>`_ for more information.
+
+.. autoclass:: pyhealth.datasets.SHHSECGDataset
+    :members:
+    :undoc-members:
+    :show-inheritance:
