--- docs/api/tasks/pyhealth.tasks.SleepStagingSHHS.rst
@@ -1,7 +1,0 @@
-pyhealth.tasks.SleepStagingSHHS
-=======================================
-
-.. autoclass:: pyhealth.tasks.shhs_sleep_staging.SleepStagingSHHS
-    :members:
-    :undoc-members:
-    :show-inheritance:
