+++ docs/api/tasks/pyhealth.tasks.SleepStagingSHHSIBI.rst
@@ -1,0 +1,7 @@
+pyhealth.tasks.SleepStagingSHHSIBI
+=======================================
+
+.. autoclass:: pyhealth.tasks.shhs_ibi_sleep_staging.SleepStagingSHHSIBI
+    :members:
+    :undoc-members:
+    :show-inheritance:
