+++ docs/api/tasks/pyhealth.tasks.dreamt_sleep_staging.rst
@@ -1,0 +1,7 @@
+pyhealth.tasks.dreampt_sleep_staging
+====================================
+
+.. autoclass:: pyhealth.tasks.dreamt_sleep_staging.SleepStagingDREAMT
+    :members:
+    :undoc-members:
+    :show-inheritance:
