+++ pyhealth/tasks/__init__.py
@@ -56,7 +56,7 @@ from .sleep_staging import (
     sleep_staging_shhs_fn,
     sleep_staging_sleepedf_fn,
 )
-from .shhs_sleep_staging import SleepStagingSHHS
+from .shhs_ibi_sleep_staging import SleepStagingSHHSIBI
 from .sleep_staging_v2 import SleepStagingSleepEDF
 from .temple_university_EEG_tasks import (
     EEGEventsTUEV,
